Fix the site
Currently the site will not render due to some mysterious javascript problem.

By trial and error I figured out that removing the google analytics code would allow the site to render again. When I replaced the existing code with new code from google, it worked.

After doing this there was still an error saying that `headsmart` was not defined. I swapped out the $().ready function call for pure JS and that got rid of that error.

No idea why that worked, but it worked.

This PR allows the site to render again.